requirements.txt
appdirs
pyyaml
PyQt
inotify
